[templates] Useful exceptions on top lvl methods
Underlying loading functions should not die when an
exception occurs, because it might be the middle
of a nested file load. If only one of the nested
files is corrupted or unavailable, the rest should
still be loaded.
Only the top level methods should raise an exception
if the loading failed in general.