[odml/property] Fix py2 odml tuple issue
When loading odml style tuples from a yaml or json file
with Python 2, the string containing the tuples is not
properly parsed from a unicode string to a list of
odml style tuples.
With JSON and YAML in Python 3 this conversion works out
of the box but to not break Python 2 compatibility already
this hotfix is introduced.