[validation] Remove default terminology check
The property_terminology_check validation can lead to a
threading lock in some cases when using Python 3.8 though
I was not able to figure out the details yet. Therefore
the validation is kept, but removed from the default
Validation checks until this can be resolved.

Also brings the docstring up to date and adjusts the
corresponding test.